Use MARISA-tries instead of dicts for dictionaries
Memory consumption of simplemma has been known to be quite high and can
easily exceed multiple GigaByte when loading the word dictionaries
for multiple languages.

To reduce memory consumption, this commit replaces the use of dicts to
store the word dictionaries with MARISA-tries, which are much more
memory efficient. This is done by utilizing [marisa-trie][1], which uses
[libmarisa][2] under the hood.

While this change massively improves loading time and memory consumption
of dictionaries, the trade-off is that the size of pickled dictionaries
increases modestly and the lookup performance becomes way worse than
before. Also pickled tries aren't architecture-independent.

[1]: https://github.com/pytries/marisa-trie
[2]: https://github.com/s-yata/marisa-trie